Improve range calculation and tick spacing and placement
Range is now chosen so that tick labels are guaranteed to be as close to the actual value they label as possible (the height of a terminal row being the limiting factor). Ticks are placed at multiples of 10^n * {1, 2, 5} so they are always round numbers. The target number of ticks and the desired buffer at the top and bottom of the range in percent can be easily adjusted (in code).